+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: paulorod <paulorod@student.42.fr>          +#+  +:+       +#+         #
+#    By: ffilipe- <ffilipe-@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/10/23 20:49:53 by paulorod          #+#    #+#              #
-#    Updated: 2023/10/26 13:43:13 by paulorod         ###   ########.fr        #
+#    Updated: 2023/10/27 15:33:15 by ffilipe-         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -22,6 +22,7 @@ SRCS =	src/main.c \
 		src/rendering/mlx_initialization.c \
 		src/rendering/render_utils.c \
 		src/rendering/rendering.c \
+		src/rendering/render_minimap.c \
 		src/utils.c
 
 MLX_PATH = ./minilibx-linux
